Move TextUnitInput.js to TextUnitInput/index.js
